requirements.txt
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mysql-connector==2.1.4
Werkzeug==0.11.11
wheel==0.24.0
